mingw: fix various warnings
 - remove unused variable
 - replace deprecated JsonReader with CharReader
 - fix initialization order
 - fix signed-unsigned comparison
 - fix string constant to char* conversion
 - fix unknown (to mingw) format character
 - fix passing NULL to non-pointer agrument
 - remove unneeded #pragma once

+++ openvpn/tun/win/tunutil.hpp
@@ -609,7 +609,7 @@ namespace openvpn {
 	{
 	  if (list)
 	    {
-	      for (unsigned int i = 0; i < list->NumAdapters; ++i)
+	      for (LONG i = 0; i < list->NumAdapters; ++i)
 		{
 		  IP_ADAPTER_INDEX_MAP* inter = &list->Adapter[i];
 		  if (index == inter->Index)
